feature: selected movie page and right movie det
