+++ src/features/movies/pages/SelectedMoviePage.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { useRouteMatch } from "react-router-dom";
+
+import MovieDetails from "../organisms/MovieDetails";
+
+export default function SelectedMoviePage() {
+  const {
+    params: { movieId },
+  } = useRouteMatch();
+
+  return <MovieDetails movieId={movieId} />;
+}
