+++ src/index.css
@@ -47,3 +47,11 @@
 .max-h-screen-96px {
   max-height: calc(100vh - 96px);
 }
+
+.max-h-32 {
+  max-height: 8rem;
+}
+
+.max-h-48 {
+  max-height: 12rem;
+}
